+++ src/Effects/RainbowFontsSL/RainbowFontsSLPS.cpp
@@ -43,7 +43,7 @@ void RainbowFontsSLPS::update(){
             //reverse the line number so that the effect moves positivly along the strip
             lineNum = numLines - i - 1;
 
-            for (uint8_t j = 0; j < numSegs; j++) {
+            for (uint16_t j = 0; j < numSegs; j++) {
                 //get the physical pixel location based on the line and seg numbers
                 //and then write out the color
                 //Note that the actual line written to is offset and wraps
